pyproject.toml
[tool.poetry]
name = "score-board-back"
version = "0.0.1"
description = "API to store game score"
authors = ["Julie Agopian <jul.ago.dev@gmail.com>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2"
djangorestframework = "^3.16"
psycopg = {extras = ["binary"], version = "*"}
gunicorn = "^23.0" 
django-cors-headers = "^4.7.0"
django-environ = "^0.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
